working on time bug

+++ app/api/art_listings.py
@@ -23,12 +23,14 @@ def get_a_single_listing(artlisting_id):
         return {"errors": "Artlisting not found."}, 404
 
     if request.method == "DELETE":
-        # if not single_listing.check_owner(owner_id):
-        #     return {"errors": "Forbidden."}, 403
-        # else:
-        db.session.delete(single_listing)
-        db.session.commit()
-        # return {"Success": "Listing deleted."}, 202
+        print("=============================",
+              owner_id, single_listing.owner_id)
+        if not single_listing.check_owner(owner_id):
+            return {"errors": "Forbidden."}, 403
+        else:
+            db.session.delete(single_listing)
+            db.session.commit()
+            return {"Success": "Listing deleted."}, 202
 
     if request.method == "PUT":
         form = ArtlistingForm()
